requirements.txt
pyttsx3
pywin32
SpeechRecognition
pyaudio
flask
gTTS
playsound
deepspeech~=0.9.3
webrtcvad~=2.0.10
halo~=0.0.18
numpy>=1.15.1
scipy>=1.1.0
TensorFlowTTS
gradio 
SoundFile
git+https://github.com/repodiac/german_transliterate
tensorflow-cpu
transformers